requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
pyyaml>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
datasets>=4.8.0
